pyproject.toml
[project]
name = "zstagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "dashscope>=1.25.15",
    "langchain>=1.2.14",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "pypdf>=6.9.2",
    "pyyaml>=6.0.3",
    "streamlit>=1.56.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
